% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_extensions.R
\name{get_USFWS_palettes}
\alias{get_USFWS_palettes}
\title{Retrieve USFWS Palettes}
\usage{
get_USFWS_palettes(palette = "dolly", reverse = FALSE, ...)
}
\description{
Easily retrieve various palettes from the USFWS collection.
}
