% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_extensions.R
\name{scale_color_USFWS}
\alias{scale_color_USFWS}
\title{Change color to one of USFWS's palettes}
\usage{
get_USFWS_palettes(palette="main", reverse=  F, ...)
}
\arguments{
\item{palette}{Name of the palette (dolly, sockeye, coaster, aurora, florida)}

\item{reverse}{Boolean: should the palette be reversed}
}
\description{
Custom color scale to use the USFWS palettes.
}
\examples{
library(tidyverse)
library(USFWSpalettes)
diamonds \%>\%
 ggplot(aes(x=cut, fill=cut))+
 geom_bar()+
 theme_minimal()+
 scale_fill_USFWS()


}
