% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_extensions.R
\name{scale_fill_USFWS}
\alias{scale_fill_USFWS}
\title{Change fill to one of USFWS's palettes}
\usage{
get_USFWS_palettes(palette="main", reverse=  F, ...)
}
\arguments{
\item{palette}{Name of the palette (dolly, sockeye, coaster, aurora, florida)}

\item{reverse}{Boolean: should the palette be reversed}
}
\description{
Custom color scale to use the USFWS palettes.
}
