% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plaettes.R
\name{USFWS_palette}
\alias{USFWS_palette}
\title{A Palette Generator}
\usage{
USFWS_palette(n = 5, name = "pal1", discrete = TRUE)
}
\arguments{
\item{n}{Number of colors}

\item{name}{Name of palette.See details for options}

\item{discrete}{Boolean value TRUE if discrete FALSE for continuous}
}
\description{
This function creates palettes that were created from photos posted by the US Fish And Wildlife Service's Instagram
}
