% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plaettes.R
\docType{data}
\name{USFWS_palettes}
\alias{USFWS_palettes}
\title{A List of All Palettes Used}
\format{An object of class \code{list} of length 5.}
\usage{
USFWS_palettes
}
\description{
A list that contains all palettes used in this package
}
\keyword{datasets}
